+++ examples/apprtc/apprtc.py
@@ -83,7 +83,7 @@ async def run(pc, player, recorder, signaling):
                 await pc.setLocalDescription(await pc.createAnswer())
                 await signaling.send(pc.localDescription)
         elif isinstance(obj, RTCIceCandidate):
-            pc.addIceCandidate(obj)
+            await pc.addIceCandidate(obj)
         elif obj is BYE:
             print("Exiting")
             break